pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "Initial setup"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.59.0",
    "duckdb>=1.3.2",
    "ipykernel>=6.30.0",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.3",
    "pymupdf4llm>=0.0.27",
    "sqlalchemy>=2.0.41",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent = ["*.json", "*.yaml", "*.yml"]
